feat(deck): Autologin to Plasma
Currently the user will boot to SDDM upon initial install. This is
not entirely optimal as the default session in SDDM is gamescope
session which is not usable without first updating Steam. This
addresses that by using the same logic we use for gamescope autologin
to login to Plasma on first boot